% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{cluster_ts}
\alias{cluster_ts}
\title{cluster_ts}
\usage{
cluster_ts(.tag_dat, .clusters)
}
\arguments{
\item{.tag_dat}{A time series with \code{date} and \code{value} cols.}

\item{.clusters}{The number of clusters.}
}
\value{
A list.
}
\description{
cluster_ts
}
