% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml-models.R
\name{coeteris_paribus}
\alias{coeteris_paribus}
\title{coeteris_paribus}
\usage{
coeteris_paribus(.model, .newdata, .target)
}
\arguments{
\item{.model}{A fitted model from \code{tidymodels}.}

\item{.newdata}{A data frame.}

\item{.target}{The name of the dependent variable.}
}
\value{
A \code{ggplot2} plot.
}
\description{
coeteris_paribus
}
