% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag-assessment.R
\name{extract_anomalies}
\alias{extract_anomalies}
\title{extract_anomalies}
\usage{
extract_anomalies(data, date_col, value_col, alpha = 0.5, anomaly = "Yes")
}
\arguments{
\item{data}{A data frame.}

\item{date_col}{A string, the name of the date column.}

\item{value_col}{A string, the name of the value column.}

\item{alpha}{Controls the width of the "normal" range regarding anomaly
detection. Lower values are more conservative while higher values are less
prone to incorrectly classifying "normal" observations.}

\item{anomaly}{A string, either "Yes" or "No".}
}
\value{
A data frame
}
\description{
extract_anomalies
}
