% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-series.R
\name{forecast_nnetar}
\alias{forecast_nnetar}
\title{forecast_nnetar}
\usage{
forecast_nnetar(.tag_dat, .assess = 1, .horiz = 1, .seasonal_period = "auto")
}
\arguments{
\item{.tag_dat}{A time series with \code{date} and \code{value} cols.}

\item{.assess}{The number of samples used for each assessment resample.}

\item{.horiz}{The forecast horizon.}

\item{.seasonal_period}{A seasonal frequency. Uses "auto" by default. A
character phrase of "auto" or time-based phrase of "2 weeks" can be used if
a date or date-time variable is provided.}
}
\value{
A \code{parsnip} (NNAR) model object.
}
\description{
forecast_nnetar
}
