% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-series.R
\name{forecast_prophet}
\alias{forecast_prophet}
\title{forecast_prophet}
\usage{
forecast_prophet(.tag_dat, .assess = 1, .horiz = 1)
}
\arguments{
\item{.tag_dat}{A time series with \code{date} and \code{value} cols.}

\item{.assess}{The number of samples used for each assessment resample.}

\item{.horiz}{The forecast horizon.}
}
\value{
A \code{parsnip} (PROPHET) model object.
}
\description{
forecast_prophet
}
