% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{forecast_tag}
\alias{forecast_tag}
\title{forecast_tag}
\usage{
forecast_tag(.tag_dat, .ndays = 15, .interactive = FALSE)
}
\arguments{
\item{.tag_dat}{A time series with 'date' (days) and 'value' cols.}

\item{.ndays}{The forecasting horizon, in days.}

\item{.interactive}{Returns either a static (ggplot2) visualization or an
interactive (plotly) visualization.}
}
\value{
Either a \code{ggplot2} or a \code{plotly} plot.
}
\description{
forecast_tag
}
