% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-series.R
\name{forecast_tbats}
\alias{forecast_tbats}
\title{forecast_tbats}
\usage{
forecast_tbats(
  .tag_dat,
  .assess = 1,
  .horiz = 1,
  .seasonal_period_1 = "auto",
  .seasonal_period_2 = NULL,
  .seasonal_period_3 = NULL
)
}
\arguments{
\item{.tag_dat}{A time series with \code{date} and \code{value} cols.}

\item{.assess}{The number of samples used for each assessment resample.}

\item{.horiz}{The forecast horizon.}

\item{.seasonal_period_1}{(required) The primary seasonal frequency. Uses
"auto" by default. A character phrase of "auto" or time-based phrase of "2
weeks" can be used if a date or date-time variable is provided.}

\item{.seasonal_period_2}{(optional) A second seasonal frequency. Is NULL
by default. A character phrase of "auto" or time-based phrase of "2 weeks"
can be used if a date or date-time variable is provided.}

\item{.seasonal_period_3}{(optional) A third seasonal frequency. Is NULL by
default. A character phrase of "auto" or time-based phrase of "2 weeks" can
be used if a date or date-time variable is provided.}
}
\value{
A \code{parsnip} (BATS) model object.
}
\description{
forecast_tbats
}
