% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{impute_missing_values}
\alias{impute_missing_values}
\title{impute_missing_values}
\usage{
impute_missing_values(.tag_dat)
}
\arguments{
\item{.tag_dat}{A time series with 'date' (days) and 'value' cols.}
}
\value{
A data frame with imputed values.
}
\description{
impute_missing_values
}
