% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nexaverser-package.R
\docType{package}
\name{nexaverser-package}
\alias{nexaverser}
\alias{nexaverser-package}
\title{nexaverser: TAG Assessment, Feature Selection and Modeling Tools}
\description{
\code{nexaverser} provides tag analysis tools for plant optimization projects or IoT systems.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/octaviodeliberato/nexaverser}
  \item Report bugs at \url{https://github.com/octaviodeliberato/nexaverser/issues}
}

}
\author{
\strong{Maintainer}: Octavio Deliberato Neto \email{octavio.deliberato@aggxtream.com}

}
\keyword{internal}
