% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml-models.R
\name{optimize_with_cobyla}
\alias{optimize_with_cobyla}
\title{optimize_with_cobyla}
\usage{
optimize_with_cobyla(
  .model,
  .vars,
  .lower,
  .upper,
  .x0,
  .maxiter = 10L,
  .option = "minimize"
)
}
\arguments{
\item{.model}{A \code{parsnip} trained model or \code{workflow}.}

\item{.vars}{A character vector with the names of the input variables.}

\item{.lower}{A vector of lower bounds for the vaiables in the function.}

\item{.upper}{A vector of upper bounds for the vaiables in the function.}

\item{.x0}{The starting point for searching the optimum.}

\item{.maxiter}{The number of iterations to run for finding a solution.}

\item{.option}{A string, either "maximize" or "minimize" the function.}
}
\value{
A \code{tibble}.
}
\description{
optimize_with_cobyla
}
