% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{plot_tag_dat}
\alias{plot_tag_dat}
\title{plot_tag_dat}
\usage{
plot_tag_dat(.tag_dat, .ncol = 2, .nrow = 1, .loess = TRUE, .plotly = FALSE)
}
\arguments{
\item{.tag_dat}{A data frame containing one date (dttm) and some numeric
columns.}

\item{.ncol}{The number of columns per page.}

\item{.nrow}{The number of rows per page.}

\item{.loess}{Logical, whether or not to include a LOESS smoother.}

\item{.plotly}{Returns either a static (ggplot2) visualization or an
interactive (plotly) visualization.}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot.
}
\description{
This is a data viz function to automagically plot the time
series.
}
\details{
This uses the \code{trelliscopejs::facet_trelliscope()} to finalize the plots.
}
\author{
Octavio Deliberato Neto.
}
