% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-viz.R
\name{plot_tag_data}
\alias{plot_tag_data}
\title{plot_tag_data}
\usage{
plot_tag_data(.tag_dat, .ncol = 2, .nrow = 1, .loess = TRUE, .plotly = FALSE)
}
\arguments{
\item{.tag_dat}{A data frame containing one date (dttm, daily data) and tag (numeric) columns.}

\item{.ncol}{The number of columns per page.}

\item{.nrow}{The number of rows per page.}

\item{.loess}{Logical - Whether or not to include a LOESS smoother.}

\item{.plotly}{Returns either a static (ggplot2) visualization or an interactive (plotly) visualization.}
}
\value{
A static \code{ggplot2} plot or an interactive \code{plotly} plot.
}
\description{
plot_tag_data
}
