% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{run_causation_analysis}
\alias{run_causation_analysis}
\title{run_causation_analysis}
\usage{
run_causation_analysis(.tag_dat, .target, .max_lag = 7, .assess = NULL)
}
\arguments{
\item{.tag_dat}{A data frame with numeric data.}

\item{.target}{The name of the target (response) column.}

\item{.max_lag}{An integer specifying the order of delays to include in the
auxiliary regression. Defaults to 7.}

\item{.assess}{The number of samples used for each test.}
}
\value{
A character vector.
}
\description{
run_causation_analysis
}
