% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-selection.R
\name{select_features_with_pps}
\alias{select_features_with_pps}
\title{select_features_with_pps}
\usage{
select_features_with_pps(
  .tag_dat,
  .target,
  .balance = FALSE,
  .return_data = FALSE,
  .task = "regression",
  .cutoff = 0.15,
  .parallel = FALSE,
  .max_cores = -1
)
}
\arguments{
\item{.tag_dat}{A time series with \code{date} (days) and \code{value} cols.}

\item{.target}{The name (string) of the target variable.}

\item{.balance}{Logical, whether or not to balance the dataset.}

\item{.return_data}{Logical, whether or not to return the dataset in the
return object.}

\item{.task}{Either "regression" or "classification".}

\item{.cutoff}{Predictors with PPS scores lower than \code{.cutoff} will be
discarded.}

\item{.parallel}{Logical, whether to perform score calls in parallel.}

\item{.max_cores}{Maximum number of cores to use, defaults to maximum
minus 1.}
}
\value{
A list.
}
\description{
select_features_with_pps
}
