% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{train_cubist_model}
\alias{train_cubist_model}
\title{train_cubist_model}
\usage{
train_cubist_model(
  .data,
  .target,
  .prop = 0.8,
  .strat = FALSE,
  .tune = TRUE,
  .grid_size = 10,
  .num_cores = 1,
  .best_metric = "rmse",
  .surrogate_model = FALSE
)
}
\arguments{
\item{.data}{A time series with 'date' (days) and 'value' cols.}

\item{.target}{The name (string) of the target variable.}

\item{.prop}{The proportion of data to be retained for modeling/analysis.}

\item{.strat}{Logical, whether or not to conduct stratified sampling by
'.target'}

\item{.tune}{Default is TRUE, this will create a tuning grid and tuned
workflow.}

\item{.grid_size}{Default is 10.}

\item{.num_cores}{Default is 1.}

\item{.best_metric}{Default is "rmse".}

\item{.surrogate_model}{Logical, whether or not to conduct surrogate
modeling, i.e., use all data to train the model.}
}
\value{
A list.
}
\description{
train_cubist_model
}
