% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine-learning.R
\name{train_xgboost_model}
\alias{train_xgboost_model}
\title{train_xgboost_model}
\usage{
train_xgboost_model(
  .data,
  .target,
  .prop = 0.8,
  .strat = FALSE,
  .tune = TRUE,
  .grid_size = 10L,
  .num_cores = 1L,
  .model_type = "regression",
  .surrogate_model = FALSE
)
}
\arguments{
\item{.data}{A data frame with \code{date} and numeric cols.}

\item{.target}{The name (string) of the target variable.}

\item{.prop}{The proportion of data to be retained for modeling/analysis.}

\item{.strat}{Logical, whether or not to conduct stratified sampling by
'.target'}

\item{.tune}{Default is TRUE, this will create a tuning grid and tuned
workflow.}

\item{.grid_size}{Default is 10.}

\item{.num_cores}{Default is 1.}

\item{.model_type}{Default is \code{regression}, can also be \code{classification.}}

\item{.surrogate_model}{Logical, whether or not to conduct surrogate
modeling, i.e., use all data to train the model.}
}
\value{
A list.
}
\description{
train_xgboost_model
}
