% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fz_data.R
\docType{data}
\name{fz_data}
\alias{fz_data}
\title{Data from 18 tags.}
\format{
A data frame with 1738 observations and 19 variables:
\describe{
\item{date}{Measurement date, yyy-mm-dd format}
\item{fz_s_0033_zn2}{Conc Zn Filtrado Secundario - S033}
\item{fz_fit_006}{Vazao de polpa do S033}
\item{fz_fit_023}{Vazao Filtro S034}
\item{fz_cons_floc_flonex}{Consumo Diario de Floculante Flonex 9026 SNF}
\item{fz_pit_012_h}{PRESSAO DE VACUO DA BOMBA K001}
\item{fz_tit_413}{Temperatura do Tanque 413}
\item{fz_ait_002_h}{Media hora PH TANQ B469}
\item{fz_fit_030}{Adicao leite de cal no tanque 469}
\item{fz_filtrabilidade_precip}{Filtrabilidade do ppFe}
\item{lx_fit_085}{Transmissor de Vazao Bomba P085}
\item{lx_b043_fe_2}{Concentracao de ferroso no ultimo tanque (B-43) Fe+2 B043}
\item{lx_b_35_fet}{Ferro Total no tanque 35}
\item{lx_li_claro_s026a}{Nivel de claro espessador S026A}
\item{lx_fit_022_h}{Vazao da purga do espessador S026A}
\item{lx_b022_h}{acidez TQ 21}
\item{lx_b021_fe_t}{Ferro Total B021}
\item{lx_b043_fet}{Ferro Total B043}
\item{lb_fz_filtros033silw_zn}{Teor de Zinco na Torta do Filtro S033}
}
}
\source{
\url{www.aggxtream.com}
}
\usage{
fz_data
}
\description{
A dataset containing PIMS data (daily averages) from 18 tags.
}
\keyword{datasets}
