% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machine-learning.R
\name{optimize_with_jaya}
\alias{optimize_with_jaya}
\title{optimize_with_jaya}
\usage{
optimize_with_jaya(
  .model,
  .vars,
  .lower,
  .upper,
  .maxiter = 10L,
  .option = "minimize",
  .seed = NULL
)
}
\arguments{
\item{.model}{A fitted model from one of the \code{nexaverser::train} functions.}

\item{.vars}{A character vector with the names of the input variables.}

\item{.lower}{A vector of lower bounds for the vaiables in the function.}

\item{.upper}{A vector of upper bounds for the vaiables in the function.}

\item{.maxiter}{The number of iterations to run for finding a solution.}

\item{.option}{A string, either "maximize" or "minimize" the function.}

\item{.seed}{An integer vector containing the random number generator state.}
}
\value{
A \code{tibble}.
}
\description{
optimize_with_jaya
}
