% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml-models.R
\name{plant_performance_map}
\alias{plant_performance_map}
\title{plant_performance_map}
\usage{
plant_performance_map(
  .model,
  .data,
  .xvar,
  .yvar,
  .zvar,
  .res = 16,
  .use_rgl = FALSE
)
}
\arguments{
\item{.model}{A \code{tidymodels} fitted model.}

\item{.data}{A data frame with the training set.}

\item{.xvar}{The first input.}

\item{.yvar}{The second input.}

\item{.zvar}{The dependent variable.}

\item{.res}{3D plot resolution.}

\item{.use_rgl}{Boolean. If TRUE, a 3D plot is produced by the \code{{rgl}}
package.}
}
\value{
A \code{plotly} plot.
}
\description{
plant_performance_map
}
