% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-selection.R
\name{select_features_with_trex}
\alias{select_features_with_trex}
\title{select_features_with_trex}
\usage{
select_features_with_trex(
  .tag_dat,
  .target,
  .balance = FALSE,
  .return_data = FALSE,
  .task = "regression",
  .corr_max = 0.5,
  .parallel = FALSE,
  .max_cores = min(20, max(1, parallel::detectCores(logical = FALSE) - 1))
)
}
\arguments{
\item{.tag_dat}{A time series with \code{date} (days) and \code{value} cols.}

\item{.target}{The name (string) of the target variable.}

\item{.balance}{Logical, whether or not to balance the dataset.}

\item{.return_data}{Logical, whether or not to return the dataset in the
return object.}

\item{.task}{Either "regression" or "classification".}

\item{.corr_max}{Maximum allowed correlation between any two predictors
from different clusters. Defaults to 0.5.}

\item{.parallel}{Logical. If TRUE random experiments are executed in
parallel. Defaults to FALSE.}

\item{.max_cores}{Maximum number of cores to be used for parallel processing
(default: min of 20 random experiments and number of physical cores).}
}
\value{
A list.
}
\description{
select_features_with_trex
}
